+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 2023 Nov 6
 
-* Switched from PyInstaller to Nuitka for MS Windows.  
+* Added Nuitka build for MS Windows in addition to pyinstaller.  
 
 
 ## 1.0.37